refactor animation component
